requirements.txt
# List your project dependencies here
dotenv
pytest
streamlit 
openai 
google-generativeai 
transformers 
torch
ollama
